+++ src/components/About.vue
@@ -1,81 +1,142 @@
 <template>
-    <div id="about">
+  <div id="about">
     <div class="header-container">
       <div class="usa-prose">
-        <h3 class="title-text">
-          {{ $route.meta.title }} {{ developmentTier }}
-        </h3>
+        <h1 class="title-text">
+          {{ pageTitle }} {{ title }} {{ developmentTier }}
+        </h1>
       </div>
     </div>
 
     <p>
-    <router-link to="/">The National Integrated Water Availability Assessments map</router-link> shows the latest available daily estimates of natural water storage for approximately 110,000 regions across the conterminous U.S.
-    Map shading indicates the current natural water storage relative to historical conditions for this time of year.
+      <router-link to="/">
+        The National Integrated Water Availability Assessments map
+      </router-link> shows the latest available daily estimates of natural water storage for approximately 110,000 regions across the conterminous U.S.
+      Map shading indicates the current natural water storage relative to historical conditions for this time of year.
     </p>
 
     <p>
-    Estimates of natural water storage are calculated using the National Hydrologic Model (NHM) Infrastructure (Regan et al. 2019)
-    configured with the <a href="https://www.usgs.gov/software/precipitation-runoff-modeling-system-prms">Precipitation Runoff Modeling
-    System</a> (PRMS; Markstrom et al., 2015). The PRMS is a modular, deterministic, distributed-parameter, physical process-based
-    hydrologic simulation code that can be used to evaluate the effects of various combinations of climate and landscape on
-    hydrologic response at the watershed scale  (Regan et al., 2018).The PRMS application of the NHM (NHM-PRMS) is used here to
-    represent the daily water balance across the diverse range of landscapes of the conterminous U.S. Further information on the
-    NHM Infrastructure, the PRMS model, and the NHM-PRMS application can be found in the following citations.
+      Estimates of natural water storage are calculated using the National Hydrologic Model (NHM) Infrastructure (Regan et al. 2019)
+      configured with the <a
+        href="https://www.usgs.gov/software/precipitation-runoff-modeling-system-prms"
+        @click="runGoogleAnalytics('about page', 'click', 'clicked text link for precipitation-runoff-modeling-system-prms')"
+      >Precipitation Runoff Modeling
+        System</a> (PRMS; Markstrom et al., 2015). The PRMS is a modular, deterministic, distributed-parameter, physical process-based
+      hydrologic simulation code that can be used to evaluate the effects of various combinations of climate and landscape on
+      hydrologic response at the watershed scale  (Regan et al., 2018).The PRMS application of the NHM (NHM-PRMS) is used here to
+      represent the daily water balance across the diverse range of landscapes of the conterminous U.S. Further information on the
+      NHM Infrastructure, the PRMS model, and the NHM-PRMS application can be found in the following citations.
     </p>
 
     <p>
-    Abatzoglou, J. T. (2013), <a href="http://onlinelibrary.wiley.com/doi/10.1002/joc.3413/full">Development of gridded surface meteorological data for ecological applications and modelling</a>. Int. J. Climatol., 33: 121–131.
+      Abatzoglou, J. T. (2013), <a href="http://onlinelibrary.wiley.com/doi/10.1002/joc.3413/full">Development of gridded surface meteorological data for ecological applications and modelling</a>. Int. J. Climatol., 33: 121–131.
     </p>
 
     <p>
-    Markstrom, S.L., Regan, R.S., Hay, L.E., Viger, R.J., Webb, R.M.T., Payn, R.A., and LaFontaine, J.H., 2015, PRMS-IV, the precipitation-runoff modeling system, version 4: U.S. Geological Survey Techniques and Methods, book 6, chap. B7, 158 p., <a href="http://dx.doi.org/10.3133/tm6B7">http://dx.doi.org/10.3133/tm6B7</a>.
+      Markstrom, S.L., Regan, R.S., Hay, L.E., Viger, R.J., Webb, R.M.T., Payn, R.A., and LaFontaine, J.H., 2015, PRMS-IV, the precipitation-runoff modeling system, version 4: U.S. Geological Survey Techniques and Methods, book 6, chap. B7, 158 p., <a href="http://dx.doi.org/10.3133/tm6B7">http://dx.doi.org/10.3133/tm6B7</a>.
     </p>
 
     <p>
-    Regan, R.S., Juracek, K.E., Hay, L.E., Markstrom, S.L., Viger, R.J., Driscoll, J.M., LaFontaine, J.H., and Norton, P.A., 2019, The US Geological Survey National Hydrologic Model infrastructure: Rationale, description, and application of a watershed-scale model for the conterminous United States: Environmental Modelling & Software, v. 111, p. 192-203, <a href="https://doi.org/10.1016/j.envsoft.2018.09.023">https://doi.org/10.1016/j.envsoft.2018.09.023</a>.
+      Regan, R.S., Juracek, K.E., Hay, L.E., Markstrom, S.L., Viger, R.J., Driscoll, J.M., LaFontaine, J.H., and Norton, P.A., 2019, The US Geological Survey National Hydrologic Model infrastructure: Rationale, description, and application of a watershed-scale model for the conterminous United States: Environmental Modelling & Software, v. 111, p. 192-203, <a href="https://doi.org/10.1016/j.envsoft.2018.09.023">https://doi.org/10.1016/j.envsoft.2018.09.023</a>.
     </p>
 
     <p>
-    Regan, R.S., Markstrom, S.L., Hay, L.E., Viger, R.J., Norton, P.A., Driscoll, J.M., LaFontaine, J.H., 2018, Description of the National Hydrologic Model for use with the Precipitation-Runoff Modeling System (PRMS): U.S. Geological Survey Techniques and Methods, book 6, chap B9, 38 p., <a href="https://doi.org/10.3133/tm6B9">https://doi.org/10.3133/tm6B9</a>.
+      Regan, R.S., Markstrom, S.L., Hay, L.E., Viger, R.J., Norton, P.A., Driscoll, J.M., LaFontaine, J.H., 2018, Description of the National Hydrologic Model for use with the Precipitation-Runoff Modeling System (PRMS): U.S. Geological Survey Techniques and Methods, book 6, chap B9, 38 p., <a href="https://doi.org/10.3133/tm6B9">https://doi.org/10.3133/tm6B9</a>.
     </p>
 
     <p>
-    Viger, R.J., and Bock, Andrew, 2014, GIS features of the geospatial fabric for national hydrologic modeling: U.S. Geological Survey data release, <a href="https://dx.doi.org/doi:10.5066/F7542KMD">https://dx.doi.org/doi:10.5066/F7542KMD</a>.
+      Viger, R.J., and Bock, Andrew, 2014, GIS features of the geospatial fabric for national hydrologic modeling: U.S. Geological Survey data release, <a href="https://dx.doi.org/doi:10.5066/F7542KMD">https://dx.doi.org/doi:10.5066/F7542KMD</a>.
     </p>
 
-    <h3>About IWAAs</h3>
+    <h2>About IWAAs</h2>
 
     <p>
-    The <a href="https://www.usgs.gov/mission-areas/water-resources/science/integrated-water-availability-assessments-iwaas">USGS Integrated Water Availability Assessments (IWAAs)</a>
-    are a multi-extent, stakeholder driven, near real-time census and seasonal
-    prediction of water availability for both human and ecological uses at
-    regional and national extents.
+      The <a
+        href="https://www.usgs.gov/mission-areas/water-resources/science/integrated-water-availability-assessments-iwaas"
+        @click="runGoogleAnalytics('about page', 'click', 'clicked text link for IWAAs')"
+      >USGS Integrated Water Availability Assessments (IWAAs)</a>
+      are a multi-extent, stakeholder driven, near real-time census and seasonal
+      prediction of water availability for both human and ecological uses at
+      regional and national extents.
     </p>
 
     <p>
-    The USGS IWAAs combine resources and knowledge gained from previous and
-    ongoing USGS efforts such as <a href="https://water.usgs.gov/watercensus/focusarea.html">Focus Area Studies</a>,
-    <a href="https://water.usgs.gov/watercensus/groundwater.html">Regional Groundwater Availability Studies</a>,
-    <a href="https://water.usgs.gov/watercensus/water-use.html">Water Use estimation</a>,
-    <a href="https://water.usgs.gov/watercensus/ecowater.html">Ecological Water Needs</a>,
-    and <a href="https://water.usgs.gov/watercensus/streamflow.html">Streamflow estimation</a>,
-    in addition to utilizing and providing feedback to the
-    <a href="http://water.noaa.gov/about/nwm">National Water Model</a>. The IWAAs
-    are designed to provide information to meet the goals of the
-    <a href="https://water.usgs.gov/watercensus/index.html">National Water Census</a> as established through the SECURE Water Act.
+      The USGS IWAAs combine resources and knowledge gained from previous and
+      ongoing USGS efforts such as <a
+        href="https://water.usgs.gov/watercensus/focusarea.html"
+        @click="runGoogleAnalytics('about page', 'click', 'clicked text link for Focus Area Studies')"
+      >Focus Area Studies</a>,
+      <a
+        href="https://water.usgs.gov/watercensus/groundwater.html"
+        @click="runGoogleAnalytics('about page', 'click', 'clicked text link for Regional Groundwater Availability Studies')"
+      >Regional Groundwater Availability Studies</a>,
+      <a
+        href="https://water.usgs.gov/watercensus/water-use.html"
+        @click="runGoogleAnalytics('about page', 'click', 'clicked text link for Water Use Estimation')"
+      >Water Use estimation</a>,
+      <a
+        href="https://water.usgs.gov/watercensus/ecowater.html"
+        @click="runGoogleAnalytics('about page', 'click', 'clicked text link for Ecological Water Needs')"
+      >Ecological Water Needs</a>,
+      and <a
+        href="https://water.usgs.gov/watercensus/streamflow.html"
+        @click="runGoogleAnalytics('about page', 'click', 'clicked text link for Streamflow estimation')"
+      >Streamflow estimation</a>,
+      in addition to utilizing and providing feedback to the
+      <a
+        href="http://water.noaa.gov/about/nwm"
+        @click="runGoogleAnalytics('about page', 'click', 'clicked text link for National Water Model')"
+      >National Water Model</a>. The IWAAs
+      are designed to provide information to meet the goals of the
+      <a
+        href="https://water.usgs.gov/watercensus/index.html"
+        @click="runGoogleAnalytics('about page', 'click', 'clicked text link for National Water Census')"
+      >National Water Census</a> as established through the SECURE Water Act.
     </p>
 
-    <p><router-link to="/"><button>Return to map</button></router-link></p>
+    <p>
+      <router-link to="/">
+        <button @click="runGoogleAnalytics('about page', 'click', 'user clicked returned to map button')">
+          Return to map
+        </button>
+      </router-link>
+    </p>
   </div>
 </template>
 
 <script>
     export default {
-        name: 'About'
+        name: 'About',
+        props: {
+            title: {
+                type: String,
+                default: process.env.VUE_APP_TITLE
+            },
+            developmentTier: {
+                type: String,
+                default: process.env.VUE_APP_TIER
+            }
+        },
+        data() {
+            return {
+                pageTitle: 'About'
+            };
+        },
+        methods: {
+            runGoogleAnalytics(eventName, action, label) {
+                this.$ga.event(eventName, action, label)
+            }
+        }
     }
 </script>
 
-<style lang="scss">
+<style scoped lang="scss">
+  h1 {
+    font-size: 1.2rem;
+  }
+  h2 {
+    font-size: 1rem;
+  }
   #about {
     margin: 1em auto;
     width: 85%;